Move gds_derivatives to a better place.
